Change the behaviour of ERMDDeleteButton and ERMDRemoveRelatedButton to ensure a deleted or removed object is not still stored under the D2W context's "object" key.
Normally there will be no "object" set at this point. Alas, when removing or deleting the last object from the list of related objects, that object does end up as the D2W context's "object".
In ERMODEditRelationshipPage, when skipping the query page for related objects to move straight to the select list, the entity to use for the list is determined via ERDDefaultConfigurationNameAssignment.entityNameForContext(). If the D2W context's "object" key is not null, that EO's entity will be used. This will cause an exception if the EO is a sub-entity and EOs of other entities are included in the list, which lack one or more of the keys in "displayPropertyKeys".